Move private methods into ProjectRepo
